Moved the JHipsterApplicationExporter to a dedicated folder
