+++ jdl/jdl-importer.js
@@ -20,7 +20,7 @@ const JDLReader = require('./readers/jdl-reader');
 const DocumentParser = require('./converters/parsed-jdl-to-jdl-object/parsed-jdl-to-jdl-object-converter');
 const JDLWithoutApplicationToJSONConverter = require('./converters/jdl-to-json/jdl-without-application-to-json-converter');
 const JDLWithApplicationsToJSONConverter = require('./converters/jdl-to-json/jdl-with-applications-to-json-converter');
-const JHipsterApplicationExporter = require('./exporters/jhipster-application-exporter');
+const JHipsterApplicationExporter = require('./exporters/applications/jhipster-application-exporter');
 const JHipsterDeploymentExporter = require('./exporters/jhipster-deployment-exporter');
 const JHipsterEntityExporter = require('./exporters/jhipster-entity-exporter');
 const BusinessErrorChecker = require('./validators/business-error-checker');
